requirements.txt
gradio
requests
langchain
langchain-community
langchain_openai
langchain-core
langchain-google-genai
langchain-huggingface
langchain-groq
langchain-tavily
langchain-chroma
langgraph
huggingface_hub
arxiv
pymupdf
wikipedia
pgvector
python-dotenv
pinecone-client